clean_requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
asttokens
attrs==25.3.0
colorama
comm
coverage==7.9.1
debugpy
decorator
exceptiongroup
executing
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.73.1
idna==3.10
importlib_metadata
iniconfig==2.1.0
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2==3.1.4
jupyter_client
jupyter_core
kornia==0.8.1
kornia_rs==0.1.9
lightning==2.5.2
lightning-utilities==0.14.3
LTNtorch==1.0.2
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib-inline
mpmath==1.3.0
multidict==6.6.3
nest_asyncio
networkx==3.3
numpy==2.1.2
packaging
parso
pickleshare
pillow==11.0.0
platformdirs
pluggy==1.6.0
prompt_toolkit
propcache==0.3.2
protobuf==6.31.1
psutil
pure_eval
Pygments
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil
#pytorch-lightning==2.5.2
pywin32==308
PyYAML==6.0.2
pyzmq
setuptools==78.1.1
six
stack_data
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
#torch==2.7.1+cu118
#torchaudio==2.7.1+cu118
torchmetrics==1.7.3
#torchvision==0.22.1+cu118
tornado
tqdm==4.67.1
traitlets
typing_extensions==4.12.2
wcwidth
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zipp
